requirements.txt
Django>=1.11
lxml==4.0.0
path.py==10.4
python3-saml==1.3.0
requests==2.18.4
xmlsec==1.3.3
